feat(backend): :tada: created entities
This commit adds the `Book` and `User` interfaces to the `backend-library-inventory` project. These interfaces define the structure and properties of a book and a user respectively. Additionally, the `NewBook` and `NewUser` types are defined as subsets of the original interfaces, omitting the `id` property. This will be useful for creating new book and user objects without specifying an ID.