feat(database): :construction: Add Sequelize database configuration and models
This commit adds the Sequelize configuration file and models for the `backend-library-inventory` project. The configuration file sets up the connection to the PostgreSQL database using environment variables. The models include `UserModel` and `BookModel`, which define the structure and properties of a user and a book respectively. These changes enable the use of Sequelize for database operations and improve the organization and functionality of the backend.