feat(frontend): :memo: Add book and user types
adds the `Book` and `User` types to the frontend library inventory project. The `Book` type includes properties such as `id`, `title`, `description`, `image`, and `status`. The `User` type includes properties such as `id`, `userName`, `password`, and `image`. Additionally, the commit includes the `CreateBook`, `CreateUser`, and `LoggedUser` types, which are derived from the `Book` and `User` types with certain properties omitted.