feat(frontend): :sparkles: Update Book component to include details view
Book component in the frontend library inventory project to include a details view. The details view is controlled by the `detailsView` prop, which determines whether to display additional information and a "View Details" button. This change enhances the functionality of the Book component and improves the user experience by allowing users to view more information about a book when desired.